kernel: move xip code to dedicated file
No functional changes, just moving the XIP code out and build it only
when XIP is enabled.
